+++ src/main/kotlin/blackjack/domain/gamestate/Hit.kt
@@ -1,7 +1,7 @@
-package blackjack.gamestate
+package blackjack.domain.gamestate
 
-import blackjack.Card
-import blackjack.Cards
+import blackjack.domain.Card
+import blackjack.domain.Cards
 
 class Hit(
     val cards: Cards,
