+++ src/test/kotlin/blackjack/domain/CardDeckTest.kt
@@ -1,6 +1,6 @@
-package blackjack
+package blackjack.domain
 
-import blackjack.CardTest.Companion.SPADE_ACE
+import blackjack.domain.CardTest.Companion.SPADE_ACE
 import io.kotest.assertions.throwables.shouldThrowExactly
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
